test/inspector/runtime/evaluate-async-expected.txt

Tests that Runtime.evaluate works with awaitPromise flag.

Running test: testResolvedPromise
{
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : 0
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : (scriptId)
        stackTrace : {
            callFrames : [
            ]
        }
        text : Uncaught (in promise)
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}

Running test: testPrimitiveValueInsteadOfPromise
{
    code : -32000
    message : Result of the evaluation is not a promise
}

Running test: testObjectInsteadOfPromise
{
    code : -32000
    message : Result of the evaluation is not a promise
}

Running test: testPendingPromise
{
    result : {
        type : object
        value : {
            a : 239
        }
    }
}

Running test: testExceptionInEvaluate
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            description : 239
            objectId : 0
            type : number
            value : 239
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : (scriptId)
        text : Uncaught
    }
    result : {
        description : 239
        type : number
        value : 239
    }
}
